+++ src/AutopilotContact.php
@@ -3,6 +3,7 @@
 namespace Autopilot;
 
 use JsonSerializable;
+use Autopilot\Exceptions\AutopilotException;
 
 class AutopilotContact implements JsonSerializable
 {
